+++ server/test/src/lib/SymbolIndex.test.ts
@@ -1,0 +1,24 @@
+import { SymbolIndex } from '../../../src/lib/SymbolIndex';
+import { expect } from 'chai';
+import 'mocha';
+
+describe('SymbolIndex()', () => {
+  describe('#indexFiles()', () => {
+    it('works', () => {
+
+    });
+  });
+
+  describe('#indexFile()', () => {
+    it('works', () => {
+
+    });
+  });
+
+  describe('#find()', () => {
+    it('returns a list of SymbolIndex containing symbols', () => {
+
+    });
+  });
+})
+
